refactor: Simplify Handler
